# YJIT: Stack temp register allocation for arm64