--- a/.github/workflows/yjit-ubuntu.yml
+++ b/.github/workflows/yjit-ubuntu.yml
@@ -78,10 +78,6 @@ jobs:
             configure: "--enable-yjit=dev"
             yjit_opts: "--yjit-call-threshold=1 --yjit-verify-ctx"
 
-          - test_task: "check"
-            configure: "--enable-yjit=dev"
-            yjit_opts: "--yjit-call-threshold=1 --yjit-temp-regs=5"
-
           - test_task: "test-all TESTS=--repeat-count=2"
             configure: "--enable-yjit=dev"