--- a/yjit/src/utils.rs
+++ b/yjit/src/utils.rs
@@ -1,6 +1,7 @@
 #![allow(dead_code)] // Some functions for print debugging in here
 
 use crate::backend::ir::*;
+use crate::core::Context;
 use crate::cruby::*;
 use std::slice;
 
@@ -141,7 +142,7 @@ macro_rules! c_callable {
 }
 pub(crate) use c_callable;
 
-pub fn print_int(asm: &mut Assembler, opnd: Opnd) {
+pub fn print_int(asm: &mut Assembler, ctx: &Context, opnd: Opnd) {
     c_callable!{
         fn print_int_fn(val: i64) {
             println!("{}", val);
@@ -164,11 +165,11 @@ pub fn print_int(asm: &mut Assembler, opnd: Opnd) {
     };
 
     asm.ccall(print_int_fn as *const u8, vec![argument]);
-    asm.cpop_all();
+    asm.cpop_all(ctx);
 }
 
 /// Generate code to print a pointer
-pub fn print_ptr(asm: &mut Assembler, opnd: Opnd) {
+pub fn print_ptr(asm: &mut Assembler, ctx: &Context, opnd: Opnd) {
     c_callable!{
         fn print_ptr_fn(ptr: *const u8) {
             println!("{:p}", ptr);
@@ -179,11 +180,11 @@ pub fn print_ptr(asm: &mut Assembler, opnd: Opnd) {
 
     asm.cpush_all();
     asm.ccall(print_ptr_fn as *const u8, vec![opnd]);
-    asm.cpop_all();
+    asm.cpop_all(ctx);
 }
 
 /// Generate code to print a value
-pub fn print_value(asm: &mut Assembler, opnd: Opnd) {
+pub fn print_value(asm: &mut Assembler, ctx: &Context, opnd: Opnd) {
     c_callable!{
         fn print_value_fn(val: VALUE) {
             unsafe { rb_obj_info_dump(val) }
@@ -194,11 +195,11 @@ pub fn print_value(asm: &mut Assembler, opnd: Opnd) {
 
     asm.cpush_all();
     asm.ccall(print_value_fn as *const u8, vec![opnd]);
-    asm.cpop_all();
+    asm.cpop_all(ctx);
 }
 
 /// Generate code to print constant string to stdout
-pub fn print_str(asm: &mut Assembler, str: &str) {
+pub fn print_str(asm: &mut Assembler, ctx: &Context, str: &str) {
     c_callable!{
         fn print_str_cfun(ptr: *const u8, num_bytes: usize) {
             unsafe {
@@ -222,7 +223,7 @@ pub fn print_str(asm: &mut Assembler, str: &str) {
     let opnd = asm.lea_label(string_data);
     asm.ccall(print_str_cfun as *const u8, vec![opnd, Opnd::UImm(str.len() as u64)]);
 
-    asm.cpop_all();
+    asm.cpop_all(ctx);
 }
 
 #[cfg(test)]
@@ -262,7 +263,7 @@ mod tests {
         let mut asm = Assembler::new();
         let mut cb = CodeBlock::new_dummy(1024);
 
-        print_int(&mut asm, Opnd::Imm(42));
+        print_int(&mut asm, &Context::default(), Opnd::Imm(42));
         asm.compile(&mut cb);
     }
 
@@ -271,7 +272,7 @@ mod tests {
         let mut asm = Assembler::new();
         let mut cb = CodeBlock::new_dummy(1024);
 
-        print_str(&mut asm, "Hello, world!");
+        print_str(&mut asm, &Context::default(), "Hello, world!");
         asm.compile(&mut cb);
     }
 }