Following up https://github.com/ruby/ruby/pull/7651, this PR adds arm64 support for stack temp register allocation. Because benchmark results are generally better in both x86_64 and arm64, this PR also changes the default of `--yjit-temp-regs` to `5`.

## Benchmark
The speedups on M1 seem less dramatic compared to my Linux x86_64 environment (maybe because M1 is faster on memory access?). But still benchmark results seem generally better than not enabling it.

```
regs=0: ruby 3.3.0dev (2023-04-04T23:36:32Z yjit-stack-arm64 ca7446cd37) +YJIT [arm64-darwin22]
regs=5: ruby 3.3.0dev (2023-04-04T23:36:32Z yjit-stack-arm64 ca7446cd37) +YJIT [arm64-darwin22]

--------------  -----------  ----------  -----------  ----------  -------------  --------------
bench           regs=0 (ms)  stddev (%)  regs=5 (ms)  stddev (%)  regs=0/regs=5  regs=5 1st itr
activerecord    17.7         2.3         17.5         2.2         1.01           0.89
erubi_rails     6.1          10.6        6.1          9.8         1.00           0.65
hexapdf         922.7        1.9         916.6        2.1         1.01           0.97
liquid-c        24.5         3.6         24.4         3.6         1.01           0.82
liquid-render   49.0         3.3         48.4         3.6         1.01           0.86
mail            57.1         2.2         55.3         1.7         1.03           0.87
psych-load      888.2        0.4         880.4        0.5         1.01           1.01
railsbench      646.9        1.6         647.6        1.7         1.00           0.92
ruby-lsp        32.0         26.4        32.2         34.7        1.00           1.02
sequel          30.3         1.6         30.4         1.4         0.99           1.00
binarytrees     112.4        2.6         110.2        2.6         1.02           1.02
chunky_png      370.1        0.3         362.7        0.4         1.02           1.00
erubi           121.4        1.5         121.8        2.1         1.00           1.00
etanni          210.6        0.9         207.3        0.9         1.02           1.02
fannkuchredux   446.0        0.4         401.5        0.2         1.11           1.00
lee             482.6        0.7         474.6        0.9         1.02           1.48
nbody           41.3         0.7         42.3         0.7         0.98           0.96
optcarrot       1434.6       0.6         1364.9       0.6         1.05           1.03
ruby-json       1496.1       0.4         1506.0       0.4         0.99           0.99
rubykon         3413.4       0.4         3196.5       0.4         1.07           1.03
30k_ifelse      350.3        0.8         368.8        2.2         0.95           0.61
30k_methods     837.5        0.6         836.0        0.9         1.00           0.88
cfunc_itself    20.9         0.9         20.6         1.4         1.02           1.03
fib             34.4         0.6         32.4         1.0         1.06           1.07
getivar         40.3         51.7        27.9         71.4        1.44           1.00
keyword_args    30.4         0.8         27.4         0.9         1.11           1.12
respond_to      19.2         0.9         18.6         0.8         1.03           1.02
setivar         10.4         78.1        8.2          91.6        1.27           1.00
setivar_object  28.7         48.4        27.6         50.8        1.04           0.99
setivar_young   28.8         49.2        27.5         50.0        1.05           1.02
str_concat      23.0         2.9         22.3         3.2         1.03           1.01
throw           11.9         0.9         11.9         1.0         1.00           1.00
--------------  -----------  ----------  -----------  ----------  -------------  --------------
```

## Code size
The following stats are measured on railsbench. Unlike linux-x86_64, `inline_code_size` seems to be increased as well (maybe due to the difference in encoding?). But the total `code_region_size` increase doesn't seem to be too bad.

### Before (--yjit-temp-regs=0)
```
inline_code_size:          4,238,900
outlined_code_size:        2,153,884
code_region_size:          7,667,712
```

### After (--yjit-temp-regs=5)
```
inline_code_size:          4,338,116
outlined_code_size:        3,075,936
code_region_size:          7,684,096
```