@maximecb 2023-04-05
> The speedups on M1 seem less dramatic compared to my Linux x86_64 environment (maybe because M1 is faster on memory access?). But still benchmark results seem generally better than not enabling it.

Well, the inline code size seems to have grown, which is the opposite of x86.

It's also not what I would have expected. x86 can directly operate on memory operands, whereas arm can't. You would expect that, on arm, the code size savings would be bigger, because there are some loads that we have to insert on arm when we refer to memory operands, which we would not need anymore.

Maybe worth doing a bit more investigation and looking at the resulting machine code. See if there are redundant moves and memory operations, or some kind of bug?