@k0kubun 2023-04-05
> Maybe worth doing a bit more investigation and looking at the resulting machine code. See if there are redundant moves and memory operations, or some kind of bug?

When we're looking at stats on railsbench, I guess we have to account for the fact that the behavior of railsbench is not that deterministic compared to other smaller benchmarks. Here's the stats of a run with `--yjit-temp-regs=0` on arm64:

```
inline_code_size:          4,372,572
outlined_code_size:        2,153,544
code_region_size:          7,798,784
```

and here's the stats of a run with `--yjit-temp-regs=5` on arm64:

```
inline_code_size:          4,337,620
outlined_code_size:        3,063,208
code_region_size:          7,815,168
```

so `inline_code_size` became slightly smaller in this instance. In a small benchmark like `fib`, `inline_code_size` is always smaller (14,600 → 14,468).

I also looked at `liquid-c`, but it was as random as railsbench. `--yjit-temp-regs=5` is sometimes worse, and sometimes significantly better.

---

On railsbench, I looked at some code with registers assigned to stack temps. This is an example scenario with smaller code:

### --yjit-temp-regs=0
92 bytes

```asm
  # Block: <=>@/opt/rubies/ruby/lib/ruby/3.3.0+0/rubygems/version.rb:373 (ISEQ offset: 150, chain_depth: 1)
  # Insn: opt_eq (stack_size: 2)
  # guard arg0 fixnum
  0x107d3d86c: ldur x11, [x21]
  0x107d3d870: tst x11, #1
  0x107d3d874: b.eq #0x107d3f318
  0x107d3d878: nop
  0x107d3d87c: nop
  0x107d3d880: nop
  0x107d3d884: nop
  0x107d3d888: nop
  # guard arg1 fixnum
  0x107d3d88c: ldur x11, [x21, #8]
  0x107d3d890: tst x11, #1
  0x107d3d894: b.eq #0x107d3f334
  0x107d3d898: nop
  0x107d3d89c: nop
  0x107d3d8a0: nop
  0x107d3d8a4: nop
  0x107d3d8a8: nop
  0x107d3d8ac: ldur x11, [x21]
  0x107d3d8b0: ldur x12, [x21, #8]
  0x107d3d8b4: cmp x11, x12
  0x107d3d8b8: mov x11, #0x14
  0x107d3d8bc: mov x12, #0
  0x107d3d8c0: csel x11, x11, x12, eq
  0x107d3d8c4: stur x11, [x21]
```

### --yjit-temp-regs=5
76 bytes
```asm
  # Block: <=>@/opt/rubies/ruby/lib/ruby/3.3.0+0/rubygems/version.rb:373 (ISEQ offset: 150, chain_depth: 1)
  # reg_temps: 00000000 -> 00000011
  # Insn: opt_eq (stack_size: 2)
  # guard arg0 fixnum
  0x107edc120: tst x1, #1
  0x107edc124: b.eq #0x107ede270
  0x107edc128: nop
  0x107edc12c: nop
  0x107edc130: nop
  0x107edc134: nop
  0x107edc138: nop
  # guard arg1 fixnum
  0x107edc13c: tst x9, #1
  0x107edc140: b.eq #0x107ede2ac
  0x107edc144: nop
  0x107edc148: nop
  0x107edc14c: nop
  0x107edc150: nop
  0x107edc154: nop
  0x107edc158: cmp x1, x9
  0x107edc15c: mov x11, #0x14
  0x107edc160: mov x12, #0
  0x107edc164: csel x11, x11, x12, eq
  0x107edc168: mov x1, x11
```

On the other hand, here's an example that's generated only with `--yjit-temp-regs=5` (8 bytes):

```asm 
  # Module#===
  # spill_temps: 00000011 -> 00000000
  0x107edc1ec: stur x1, [x21]
  0x107edc1f0: stur x9, [x21, #8]
```

Savings in load instructions could be offset by spills, especially on railsbench that has many method calls and C func operations. Here's other stats on railsbench:

```
temp_reg_opnd:               109,487
temp_mem_opnd:                84,168
temp_spill:                   71,843
```

The `temp_reg_opnd / temp_spill` ratio is 1.52. On `fib`, the ratio is 1.70. Even with `fib`'s ratio, the savings in the code size is not that significant. Given the fact that `temp_reg_opnd` may not always translate to a code size reduction, e.g. `stur` is just changed to `mov`, it's not that surprising to me that the code size savings in railsbench on arm64 is minimal or randomly worse.