@maximecb 2023-04-05
> When we're looking at stats on railsbench, I guess we have to account for the fact that the behavior of railsbench is not that deterministic compared to other smaller benchmarks. Here's the stats of a run with --yjit-temp-regs=0 on arm64:

Hm. Is there nondeterministic behavior in railsbench? I tried to seed the random number generator.

I see that `SecureRandom` is being used here, which might not be seeded though?
```
benchmarks/railsbench/config/initializers/content_security_policy.rb:# Rails.application.config.content_security_policy_nonce_generator = -> request { SecureRandom.base64(16) }
```

Generally speaking, we should try to remove sources of nondeterminism in the benchmarks (by seeding RNGs and not using RNGs that can't be seeded), and we should also eliminate sources of nondeterminism from YJIT if possible.

>> Given the fact that temp_reg_opnd may not always translate to a code size reduction, e.g. stur is just changed to mov, it's not that surprising to me that the code size savings in railsbench on arm64 is minimal or randomly worse.

It's a bit disappointing when compared to x86. On arm it looks like we're just increasing complexity and compilation time for not much gains (so far).

Are the `stur` instructions mostly being used for writing the output operands of instructions? If so, I wonder if we could somehow do better by looking ahead and trying to place the output value into the output register directly when allocating instruction outputs.

Otherwise, do you have some ideas for making this better? What are the next steps, in your mind?